feat: split client and server env validation
