+++ app/api/[tenant]/ast/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { getToken } from 'next-auth/jwt'
 import { prisma } from '@/lib/prisma'
-import { SERVER_ENV } from '@/lib/env'
+import { serverEnv } from '@/lib/env.server'
 import { z } from 'zod'
 
 export const dynamic = 'force-dynamic'
@@ -12,7 +12,7 @@ const bodySchema = z.object({
 })
 
 async function ensureUser(request: NextRequest, tenantSubdomain: string) {
-  const token = await getToken({ req: request, secret: SERVER_ENV.NEXTAUTH_SECRET })
+  const token = await getToken({ req: request, secret: serverEnv.NEXTAUTH_SECRET })
   if (!token?.sub) {
     return { error: NextResponse.json({ error: 'Authentication required' }, { status: 401 }) }
   }
