+++ app/api/ast/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import { getToken } from 'next-auth/jwt'
 import { prisma } from '@/lib/prisma'
 import { z } from 'zod'
-import { SERVER_ENV } from '@/lib/env'
+import { serverEnv } from '@/lib/env.server'
 import { sanitizeFormData } from './utils'
 
 export const dynamic = 'force-dynamic'
@@ -47,7 +47,7 @@ const requestSchema = z.object({
 
 export async function POST(request: NextRequest) {
   try {
-    const token = await getToken({ req: request, secret: SERVER_ENV.NEXTAUTH_SECRET })
+    const token = await getToken({ req: request, secret: serverEnv.NEXTAUTH_SECRET })
     const userId = token?.sub
     if (!userId) {
       return NextResponse.json(
@@ -116,7 +116,7 @@ export async function POST(request: NextRequest) {
 
 export async function GET(request: NextRequest) {
   try {
-    const token = await getToken({ req: request, secret: SERVER_ENV.NEXTAUTH_SECRET })
+    const token = await getToken({ req: request, secret: serverEnv.NEXTAUTH_SECRET })
     if (!token?.sub) {
       return NextResponse.json(
         { error: 'Authentication required' },
